// src/generated/models/pedido-model.ts
import type { Empleado } from './empleado-model';

export const PedidoEstatusKeyToLabel = {
  'EstatusKey0': 'Pendiente',
  'EstatusKey1': 'Entregado',
  'EstatusKey2': 'Cancelado'
} as const;
export type PedidoEstatusKey = keyof typeof PedidoEstatusKeyToLabel;

export const PedidoMotivocancelacinKeyToLabel = {
  'MotivocancelacinKey0': 'Ausencia del empleado',
  'MotivocancelacinKey1': 'Cambio de horario',
  'MotivocancelacinKey2': 'Duplicado accidental',
  'MotivocancelacinKey3': 'Cancelado por logística',
  'MotivocancelacinKey4': 'Error en la solicitud',
  'MotivocancelacinKey5': 'Otro'
} as const;
export type PedidoMotivocancelacinKey = keyof typeof PedidoMotivocancelacinKeyToLabel;

export interface Pedido {
  /**
   * @displayName Pedido
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Pedido Nombre
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  pedidonombre: string;
  /**
   * @displayName Empleado
   */
  empleado?: Pick<Empleado, 'id' | 'nombrecompleto'> & { entraobjectid?: string; fotodeperfil?: string };
  /**
   * @displayName Empleado
   */
  empleado1?: Pick<Empleado, 'id' | 'nombrecompleto'> & { entraobjectid?: string; fotodeperfil?: string };
  /**
   * @displayName Estatus
   * @validationRule Required for create/update operations
   */
  estatusKey: PedidoEstatusKey;
  /**
   * @displayName Fecha de creación
   */
  fechadecreacin?: string;
  /**
   * @displayName Fecha Entrega
   * @validationRule Required for create/update operations
   */
  fechaentrega: string;
  /**
   * @displayName Hora Entrega
   * @validationRule Required for create/update operations
   */
  horaentrega: string;
  /**
   * @displayName Cantidad
   * @validationRule Required for create/update operations
   */
  cantidad: number;
  /**
   * @displayName Notas
   * @validationRule Required for create/update operations
   */
  notas: string;
  /**
   * @displayName Motivo Cancelación
   */
  motivocancelacinKey?: PedidoMotivocancelacinKey;
  /**
   * @displayName Motivo Cancelación (Texto Libre)
   */
  motivocancelacintextolibre?: string;
}

export const _Pedido = 'Pedido' as const;